app/requirements.txt
Django==3.2.10
gunicorn==20.1.0
psycopg2-binary==2.9.1
boto3==1.20.5
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework-simplejwt==4.7.1
Pillow==8.2.0
